+++ CHANGELOG.md
@@ -6,6 +6,12 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## Version 0.7.13 - Unreleased
 
+
+### Fixed
+
+* Polygons now better respect the color parameter in detections draw on
+
+
 ### Changed
 
 * Changed default GDAL compression in `kwimage.imwrite` from RAW to DEFLATE